add cmd for provider
