+++ cmd/retrieval-provider/sample_data.json
@@ -1,0 +1,3 @@
+[
+	"bafkreieernhx5w6ymdrc5mvbxdzp7vlqk2o25r7krl4wx7crvx6fmzgqza"
+]
